Eliminate friction glitch
This change eliminates a glitch which occurred when friction was combined with IO blending and when the amount of IO blending was low. The glitch is solved by letting the amount of physical constrainment (e.g. friction) equal the amount of IO blending. (Previously, physical constrainment was always fully applied when enabled.)