Eliminate friction + translation glitch
This change eliminates a glitch which occurred when friction was combined with IO blending in a translation-enabled profile (in this case this means a profile where the input is allowed to move across space).

The problem was that when adjusting the amount of IO blending, the IO blend would move very rapidly across space, as a result of the variable constrainment amount being locked to the IO blending amount.

This change solves the problem by decoupling constrainmant amount from IO blending amount, and replacing it with another mechanism: toggling of friction at a fixed boundary (currently IO blending = 0.5).