Towards standalone application
This change takes initial steps towards running the movement AI as a standalone application, without UI, and with more flexibility e.g. as regards behaviours and numbers of inputs and outputs.

The change contains a new script applications/only_ai.py with an improv behaviour where the output seems to properly updated each time step. However, it needs to be validated e.g. with sending OSC to a UI which can render the output.