Improve timing
This change improves the timing in the standalone application. Previously, the application ignored the processing time so that the actual frame rate was always lower than the desired one. After the change, the application roughly maintains the desired frame rate if processing is not too time consuming. If processing is too time consuming, the actual frame rate drops.

Additionally, the argument --show-fps is added for monitoring of the actual frame rate (printed to the console).