Reflect translation during mirroring
This change makes the standalone applications reflect translation during mirroring, by adding --translate to the KernelPCA profile. In the profile, the translation weight is set to zero. This is a hack for including translation in the entity (which is required for mirroring to reflect translation) and at the same time reducing its impact for PCA.

Note: Requires PCA models to be re-trained.