ai_human_duet: Fix mirror-memory translation
This change fixes a translation problem which occurred in transitions between mirror and memory. The solution consists of a complete remake of how translations are handled across behaviors. It introduces a chaining mechanism which is more general than the previous solution.

Additonally, memory now proceeds as intended during interpolation between behaviors (this seems to fix a problem which has not been validated).