ai_human_duet: Improve translation during interpolation
This change aims to fix a problem which was introduced in the previous commit. The problem was that if the source state's translation changed during interpolation, then this shift would be inherited by the next state in the second half of the interpolation, which could lead to a glitch. However, neither the problem or the solution have been properly validated.