Add support for decaying learning rate
This change adds autoencoder support for an exponentially decaying learning rate. It also enables it in the deep autoencoder profile, where it causes a significant reduction in computation time and number of epochs (for a given target loss slope), with a training result that is similar to one with a constant learning rate (but perhaps with a slightly less varied output).