Solve quaternion interpolation problem
This change solves a problem with the quaternion interpolation which could cause sudden jerks. The problem was that the algorithm could switch to another "shortest path" (in SLERP terms), in the middle of an ongoing interpolation.

Example:
- Set --camera=0.180,-4.280,-0.081,-54.000,68.300 in valencia_pn.profile
- python dim_reduce.py -p valencia_pn --enable-io-blending --color-scheme=black --floor-renderer=checkerboard --io-blending=0.5 --mode=improvise --deterministic --random-seed=0 --enable-follow --start-frame=856 --stopped
- press n for next frame

The problem is solved as follows: if there is a "shortest path", it's added as a possible candidate. The candidate with the smallest angular distance to the previous quaternion (for the respective joint) is chosen.