Ensure simple OSC sender looks up host name only once
Before this change, if simple_osc_sender.OscSender was passed a host name, and if the protocol was UDP, then it would pass the host name to socket.sendto every time something was sent. This slowed down execution (especially on OSX among the tested machines).

After this change, the host name is looked up in the constructor instead and is then passed as IPv4 string to socket.sendto.