ai_human_duet: Reduce interpolation glitch
This change reduces a glitch which could occur during interpolation between dynamic quaternions e.g. IO blending. The problem was that in some situations, the interpolation would jump from a clockwise to counterclockwise solution (or vice versa) from one time step to another.

To reproduce:
- In valencia_pn.profile, set --camera=0.180,-4.280,-0.081,-54.000,68.300
- python dim_reduce.py -p valencia_pn --enable-io-blending --color-scheme=black --floor-renderer=checkerboard --io-blending=0.5 --mode=improvise --deterministic --random-seed=0 --enable-follow --stopped --start-frame=1239
- press n for next frame
- the output flips around drastically

This change introduces an optional solution by adding a --max-angular-step argument. This argument limits the angular velocity of a given dynamic interpolation. When set to a suffiently low value (e.g. 0.1), this mechanism results in a smoothing-out of sequences that previously caused a drastic glitch.

Note however that when the mechanism can also cause an undesired smoothing of desired transient or rapid movement qualities. E.g. if IO blending is constantly activated, then the mechanism will apply even when no interpolation is taking place.

This change applies the new mechanism in ai_human_duet. (The problematic case described above remains unchanged, but the solution has been validated on it.)