learn_recall_improvise: Eliminate recall glitch
This change solves a bug which caused recall to "bleed" into improv; when recall amount was 0, the output would still be affected by what the recall behavior was doing. The bug was caused by the fact that the main behavior and the recall behavior shared the same entity, and thereby the same interpolation states.

This change adds a separate recall entity to solve the problem.